creating new component for the search tags
